requirements/requirements.txt
fastapi
psycopg
uvicorn
boto3
PyJWT
python-multipart